refactor(tactics): extract useTacticsLineup controller hook
Move the lineup state machine — roster derivation, the optimistic-XI
triangle (pendingStartingXiIds + persistStartingXI + the startingXiIds
memo + clear effect), selection/swap, drag-and-drop, and the
assign/promote/demote/formation/play-style/match-role/phase handlers —
into a single useTacticsLineup hook. TacticsTab.tsx becomes a 250-line
orchestrator (from 704) wiring three hooks (lineup, filters, library) to
the command bar, player list, pitch, right panel, and focus modal.

These flows could not be split: every lineup handler calls
persistStartingXI and clearLineupSelection and reads the same roster
state. All verbatim — the five-expression startingXiIds dep array, the
dragStateRef/hoveredSlotRef mirror pattern, dragPreviewRef (returned and
attached in the parent), and each handler's specific collection
(available vs roster vs playersById) are preserved.

TacticsTab.tsx: 963 -> 250 lines.
